+++ docs/plugin/umi-ui.md
@@ -360,8 +360,11 @@ api.addPanel({
 });
 ```
 
+### api.Field
 
-## api.notify
+
+
+### api.notify
 
 调用 Umi UI 通知栏，若用户停留在当前浏览器窗口，通知栏样式为 antd [Notification](https://ant.design/components/notification-cn)，否则为系统原生通知栏。
 
@@ -403,7 +406,7 @@ notify({
 });
 ```
 
-## api.redirect
+### api.redirect
 
 项目详情内的路由跳转，在不同插件之间进行跳转。
 
@@ -421,7 +424,7 @@ export default () => (
 );
 ```
 
-## api.currentProject
+### api.currentProject
 
 获取当前项目基本信息，信息包括：
 
@@ -449,7 +452,7 @@ export default () => (
 );
 ```
 
-## api.debug
+### api.debug
 
 `debug` [API](https://github.com/visionmedia/debug#browser-support)。
 
@@ -466,12 +469,12 @@ export default () => {
 }
 ```
 
-![image](https://user-images.githubusercontent.com/13595509/63997643-aa268d00-cb31-11e9-9d42-117abd7267c9.png)
+![image](https://gw.alipayobjects.com/zos/antfincdn/0%26OBjUyZHz/01bc8117-fa3a-4421-b2f1-e8784883ffff.png)
 
 > 不建议在插件里使用 `console.log` 调用。
 
 
-## api.getCwd
+### api.getCwd
 
 获取 Umi UI 启动时的路径。
