+++ packages/umi-plugin-ui/src/plugins/configuration/locales/zh-CN.ts
@@ -1,13 +1,10 @@
 export default {
-  'org.umi.ui.configuration.object.item.delete.confirm': '确认删除？',
-  'org.umi.ui.configuration.object.item.delete.tooltip': '删除一项',
   'org.umi.ui.configuration.panel': '配置',
   'org.umi.ui.configuration.actions.open.config': '打开配置文件',
   'org.umi.ui.configuration.project.config.title': '项目配置',
   'org.umi.ui.configuration.project.config.desc': '配置 {library} 项目',
   'org.umi.ui.configuration.project.config.react.title': 'umi-plugin-react',
   'org.umi.ui.configuration.search.placeholder': '搜索',
-  'org.umi.ui.configuration.detail': '查看详情',
   'org.umi.ui.configuration.okText': '确定',
   'org.umi.ui.configuration.cancelText': '取消',
   'org.umi.ui.configuration.save': '保存',
@@ -15,9 +12,4 @@ export default {
   'org.umi.ui.configuration.reset.tooltip': '你确定要重置以上配置吗',
   'org.umi.ui.configuration.reset.tooltip.desc': '当前配置 {value} 已被修改',
   'org.umi.ui.configuration.reset.tooltip.desc.empty': '配置没有修改',
-  'org.umi.ui.configuration.add.column': '添加一列',
-  'org.umi.ui.configuration.edit.in.editor': '在编辑器中编辑',
-  'org.umi.ui.configuration.open.editor.failure': '打开编辑器失败',
-  'org.umi.ui.configuration.basic.config.object.error': '配置参数不正确',
-  'org.umi.ui.configuration.basic.config.object.select.error': '请选择相应的配置',
 };
