+++ packages/umi-ui/client/src/components/Field/Any.tsx
@@ -1,14 +1,27 @@
 import * as React from 'react';
 import { Export } from '@ant-design/icons';
 import { message, Form } from 'antd';
+import { formatMessage } from 'umi-plugin-react/locale';
 import Label from './label';
+import Context from '@/layouts/Context';
+import { callRemote } from '@/socket';
 import { ICompProps } from './index';
-import Context from '../../Context';
 import styles from './styles.module.less';
 
 const AnyComp: React.SFC<ICompProps> = props => {
+  const _log = g_uiDebug.extend('Field:AnyComp');
   const { name, description, title, link } = props;
-  const { debug: _log, openConfigAction, api } = React.useContext(Context);
+  const { currentProject } = React.useContext(Context);
+  const openConfigAction = {
+    title: 'org.umi.ui.configuration.actions.open.config',
+    type: 'default',
+    action: {
+      type: '@@actions/openConfigFile',
+      payload: {
+        projectPath: currentProject ? currentProject.path : '',
+      },
+    },
+  };
   const { action } = openConfigAction;
   const basicItem = {
     name,
@@ -18,12 +31,12 @@ const AnyComp: React.SFC<ICompProps> = props => {
 
   const handleClick = async () => {
     try {
-      await api.callRemote(action);
+      await callRemote(action);
     } catch (e) {
       message.error(
         e && e.message
           ? e.message
-          : api.intl({ id: 'org.umi.ui.configuration.open.editor.failure' }),
+          : formatMessage({ id: 'org.umi.ui.configuration.open.editor.failure' }),
       );
       _log('AnyComp error', e);
     }
@@ -32,7 +45,9 @@ const AnyComp: React.SFC<ICompProps> = props => {
     <Form.Item {...basicItem}>
       <div className={styles['any-field']}>
         <Export />
-        <a onClick={handleClick}>{api.intl({ id: 'org.umi.ui.configuration.edit.in.editor' })}</a>
+        <a onClick={handleClick}>
+          {formatMessage({ id: 'org.umi.ui.configuration.edit.in.editor' })}
+        </a>
       </div>
     </Form.Item>
   );
