+++ packages/umi-ui/client/src/components/Field/Boolean.tsx
@@ -2,18 +2,19 @@ import * as React from 'react';
 import { Form, Switch } from 'antd';
 import { ICompProps } from './index';
 import Label from './label';
-import Context from '../../Context';
 import { getFormItemShow } from './utils';
 
 const BooleanComp: React.SFC<ICompProps> = props => {
+  const _log = g_uiDebug.extend('Field:BooleanComp');
+  debugger;
   const { name, description, title, link } = props;
-  const { debug: _log } = React.useContext(Context);
   const { parentConfig } = getFormItemShow(name);
   const basicItem = {
     name,
     label: <Label name={name} title={title} description={description} link={link} />,
     valuePropName: 'checked',
   };
+
   return parentConfig ? (
     <Form.Item noStyle shouldUpdate={(prev, curr) => prev[parentConfig] !== curr[parentConfig]}>
       {({ getFieldValue }) => {
