+++ packages/umi-ui/client/src/components/Field/TextArea.tsx
@@ -1,34 +1,29 @@
 import React from 'react';
-import { Form, Select } from 'antd';
+import { Form, Input } from 'antd';
+import { formatMessage } from 'umi-plugin-react/locale';
 import { ICompProps } from './index';
 import Label from './label';
 import { getFormItemShow } from './utils';
 
-const { Option } = Select;
+const { TextArea } = Input;
 
-const ListComp: React.SFC<ICompProps> = props => {
-  const { name, description, form, title, choices, link } = props;
-  const { parentConfig } = getFormItemShow(name, form);
+const TextAreaComp: React.SFC<ICompProps> = props => {
+  const _log = g_uiDebug.extend('Field:TextAreaComp');
+  const { name, description, title, default: defaultValue, link } = props;
+  const { parentConfig } = getFormItemShow(name);
   const basicItem = {
     name,
+    required: false,
     label: <Label name={name} title={title} description={description} link={link} />,
+    rules: [{ required: !!defaultValue, message: `请输入${title}` }],
   };
 
-  const formControl = (
-    <Select style={{ maxWidth: 320 }} getPopupContainer={triggerNode => triggerNode.parentNode}>
-      {Array.isArray(choices) &&
-        choices.map(choice => (
-          <Option key={choice} value={choice}>
-            {choice}
-          </Option>
-        ))}
-    </Select>
-  );
+  const formControl = <TextArea autoComplete="off" rows={4} style={{ maxWidth: 320 }} />;
 
   return parentConfig ? (
     <Form.Item shouldUpdate={(prev, curr) => prev[parentConfig] !== curr[parentConfig]} noStyle>
       {({ getFieldValue }) => {
-        console.log(
+        _log(
           'children field update',
           name,
           parentConfig,
@@ -51,4 +46,4 @@ const ListComp: React.SFC<ICompProps> = props => {
   );
 };
 
-export default ListComp;
+export default TextAreaComp;
