+++ packages/umi-ui/client/src/components/Field/label.tsx
@@ -1,6 +1,7 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import cls from 'classnames';
-import Context from '../../Context';
+import { formatMessage } from 'umi-plugin-react/locale';
+import Context from '@/layouts/Context';
 import styles from './styles.module.less';
 
 interface ILabelProps {
@@ -13,8 +14,6 @@ interface ILabelProps {
 const Label: React.SFC<ILabelProps> = ({ name, title, description, link }) => {
   const { theme } = React.useContext(Context);
   const wrapCls = cls(styles.label, styles[`label-${theme}`]);
-  const { api } = useContext(Context);
-  const { intl } = api;
 
   return (
     <div className={wrapCls} id={name}>
@@ -23,8 +22,8 @@ const Label: React.SFC<ILabelProps> = ({ name, title, description, link }) => {
         <p>
           {description}{' '}
           {link && (
-            <a href={link} target="_blank">
-              {intl({ id: 'org.umi.ui.configuration.detail' })}
+            <a href={link} target="_blank" rel="noopener noreferrer">
+              {formatMessage({ id: 'org.umi.ui.configuration.detail' })}
             </a>
           )}
         </p>
