+++ packages/umi-ui/client/src/components/Field/utils.ts
@@ -1,0 +1,9 @@
+import { ICompProps } from './index';
+
+export const getFormItemShow = (name: string) => {
+  const configs = name.split('.');
+  const parentConfig = configs.length > 1 ? configs[0] : '';
+  return {
+    parentConfig,
+  };
+};
