+++ packages/umi-ui/client/src/locales/zh-CN.ts
@@ -85,4 +85,12 @@ export default {
   'org.umi.ui.global.copy.success': '复制路径成功',
   'org.umi.ui.global.error.title': '出错了，可点击右下角进行反馈',
   'org.umi.ui.global.error.stack': '错误堆栈',
+  'org.umi.ui.configuration.add.column': '添加一列',
+  'org.umi.ui.configuration.object.item.delete.confirm': '确认删除？',
+  'org.umi.ui.configuration.object.item.delete.tooltip': '删除一项',
+  'org.umi.ui.configuration.open.editor.failure': '打开编辑器失败',
+  'org.umi.ui.configuration.edit.in.editor': '在编辑器中编辑',
+  'org.umi.ui.configuration.detail': '查看详情',
+  'org.umi.ui.configuration.basic.config.object.error': '配置参数不正确',
+  'org.umi.ui.configuration.basic.config.object.select.error': '请选择相应的配置',
 };
